Q doesn't need to be size_t, it can be uint32_t

+++ include/params.hpp
@@ -8,10 +8,10 @@ namespace frodo_params {
 // Compile-time executable check for ensuring that FrodoKEM's parameter i.e.
 // integer modulus Q has correct value.
 constexpr bool
-check_q(const size_t q)
+check_q(const uint32_t q)
 {
-  return ((q & (q - 1)) == 0)   // q must be power of 2
-         && (q <= (1ul << 16)); // and it must be <= 2^16
+  return ((q & (q - 1)) == 0)  // q must be power of 2
+         && (q <= (1u << 16)); // and it must be <= 2^16
 }
 
 }
